Split out the xrcpreview event handlers
It's a lot of code, so it makes more sense to have them in their own file.
